tz_codec: sửa bug thật -- decodeExecuteResult làm sai EXCEPTION_NONE=-1 thành 255 qua wire
Phát hiện qua cmd/tool/tz_replay_check's differential replay dải block
thật (plan doc §9's "Giai đoạn 4"): so ModeCgo vs ModeTrustzone trên
cùng 1 dải 9 block thật (deploy+store SimpleStorage) khớp byte-for-byte
TUYỆT ĐỐI ở mọi trường (status/gas/return/state root) trừ đúng 1
trường: Exception luôn lệch -1 (A, cgo) vs 255 (B, trustzone loopback).

Root cause: pb.EXCEPTION_NONE = -1 (giá trị "thành công", chiếm gần
như mọi tx) -- mvm_tz_protocol.h's hdr.exception là uint8_t (băng
thông dây, không đổi được nếu không rebuild+reflash TA). encode
(C.uint8_t(r.Exception)) cắt -1 xuống đúng bit pattern 0xFF, ĐÚNG --
nhưng decode cũ (pb.EXCEPTION(hdr.exception)) chuyển thẳng uint8_t
(không dấu) sang pb.EXCEPTION (int32) -- Go zero-extend chứ không
sign-extend, biến 0xFF thành 255 thay vì -1. Nghĩa là MỌI tx thành
công (Status=RETURNED) qua đường trustzone/trustzone-hardware đều có
Exception() sai kể từ khi wire codec này được viết -- bug thật, không
phải giả thuyết, chỉ chưa từng bị bắt vì đây là lần đầu tiên có 1 bài
so byte-for-byte đủ chặt (dải block thật, không phải chuỗi tự dựng).

Fix: pb.EXCEPTION(int8(hdr.exception)) -- reinterpret byte thô thành
số có dấu (int8) TRƯỚC khi widen sang enum int32, đảo đúng ngược lại
phép cắt bit của encode. KHÔNG đổi layout/kích thước trường trên dây
(vẫn 1 byte) -- nên KHÔNG cần rebuild/reflash TA để có hiệu lực (chỉ
ảnh hưởng phía Go decode). Toàn bộ range thật của pb.EXCEPTION (-1..15)
nằm gọn trong int8, an toàn.

Xác nhận: rerun cùng dải 9 block sau fix -- tz_replay_check -compare-a/-b
báo "0 mismatch(es)", exit 0. Kiểm chứng ngược: cố ý dùng dải thiếu
block + tự sửa tay 1 root trong file kết quả -- cả 2 lượt đều bị báo
MISMATCH đúng, exit 1 -- xác nhận bộ so sánh thật sự phát hiện được
lệch, không phải luôn báo khớp.

Verify: go build/vet/gofmt sạch; go test -count=3 ./pkg/mvm/... sạch
(7 test loopback + 9 test dispatcher không regress); build_check.sh
ALL BUILDS PASSED (4/4).

+++ execution/pkg/mvm/tz_codec.go
@@ -368,8 +368,22 @@ func decodeExecuteResult(header, blob []byte) (*MVMExecuteResult, error) {
 	hdr := (*C.mvm_tz_execute_result_hdr_t)(unsafe.Pointer(&header[0]))
 
 	r := &MVMExecuteResult{
-		Status:    pb.RECEIPT_STATUS(hdr.status),
-		Exception: pb.EXCEPTION(hdr.exception),
+		Status: pb.RECEIPT_STATUS(hdr.status),
+		// pb.EXCEPTION's whole defined range is -1 (EXCEPTION_NONE, the
+		// overwhelmingly common "success" value) through 15 -- comfortably
+		// inside int8. hdr.exception is C.uint8_t (unsigned) purely because
+		// that's the wire's byte layout (mvm_tz_protocol.h); converting it
+		// straight to pb.EXCEPTION (int32-based) would zero-extend instead
+		// of sign-extend, turning EXCEPTION_NONE=-1 into 255 on every single
+		// successful tx. Found 2026-08-20 via cmd/tool/tz_replay_check's
+		// real-block differential replay (note/tee_dual_mode_execution_plan.md
+		// §9's "Giai đoạn 4") -- every ModeCgo-vs-ModeTrustzone receipt
+		// otherwise matched byte-for-byte except this field. int8(...) first
+		// reinterprets the raw byte as signed before widening, which is the
+		// correct inverse of encodeExecuteResult's C.uint8_t(r.Exception)
+		// truncation below -- no wire layout change, so this does NOT need
+		// a TA rebuild/reflash to take effect on the Go side.
+		Exception: pb.EXCEPTION(int8(hdr.exception)),
 		GasUsed:   uint64(hdr.gas_used),
 	}
